requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
streamlit==1.34.0
Pillow==10.3.0
plotly==5.22.0